refactor out constructDepGraph.
